ref: CourierApiRequests and OrderApiRequests renamed to CourierApiMethods and OrderApiMethods to describe better methods inside
