+++ src/test/java/com/kotan4ik/tests/order/GetListOfOrdersTest.java
@@ -4,7 +4,7 @@ import io.restassured.response.Response;
 import org.apache.http.HttpStatus;
 import org.junit.jupiter.api.Test;
 
-import static com.kotan4ik.requests.OrderApiRequests.getOrderList;
+import static com.kotan4ik.requests.OrderApiMethods.getOrderList;
 import static com.kotan4ik.utils.Assertions.compareResponseCode;
 import static com.kotan4ik.utils.Assertions.hasListOfOrdersBody;
